+++ NodeDockerfile
@@ -5,4 +5,4 @@ COPY --chown=node:node package*.json ./
 RUN npm install
 COPY --chown=node:node . .
 EXPOSE 1337
-CMD [ "node", "node.js"]
+CMD [ "node", "./app/node.js"]
